@wip
Include master variant images

When determining the images to show for a selected variant, we need to
include all images associated with the master variant as well.

STATUS: seems to work for variants, but on products the 'object' passed
to the variant scope is the Product which causes chaos and delay.